// src/actions/domain/enums.ts
/**
 * Auto generated. DO NOT edit manually.
 * Last updated on: Thu, 06 Jul 2023 15:42:04 GMT
 */

/**
 * @enum
 * @readonly
 * @description IndexWorlds Action Names
 */
export enum IndexWorldsActionName {
    Hdlegovchg = 'hdlegovchg',
    Regaccount = 'regaccount',
    Regdac = 'regdac',
    Regref = 'regref',
    Setowner = 'setowner',
    Setsociallnk = 'setsociallnk',
    Setsocials = 'setsocials',
    Setstatus = 'setstatus',
    Settitle = 'settitle',
    Unregaccount = 'unregaccount',
    Unregdac = 'unregdac',
    Unregref = 'unregref',
};